% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addExportImport.R
\name{addExportImport}
\alias{addExportImport}
\title{Export and import of areas or districts}
\usage{
addExportImport(x)
}
\arguments{
\item{x}{an object of class "antaresDataList" created with the function
\code{readAntares}. It has to contain some areas and all the links that are
connected to these areas. Moreover the function \code{\link{removeVirtualAreas}} must be call before.}
}
\description{
This function computes the export and import of areas or districts and add it to an
\code{antaresData} object.
}
\examples{
\dontrun{
mydata <- readAntares(areas = "all", links="all", districts ="all" , synthesis = FALSE)

mydata<-removeVirtualAreas(mydata, storageFlexibility = c(getAreas("psp"),getAreas("hub")), production = getAreas("off") )

mydata<-addExportImport(mydata)

names(mydata$areas)
names(mydata$districts)

}
}

