% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMonotone.R
\name{addMonotones}
\alias{addMonotones}
\title{addMonotones}
\usage{
addMonotones(antaresData = NULL)
}
\value{
\code{addMonotones} modifies its input by adding monotones.
}
\description{
This function compute monotone for some variables.
}
\examples{
\dontrun{
# First simulation
studyPath <- "path/to/study/"

setSimulationPath(studyPath, 1)
mydata1 <- readAntares(areas = "all", districts = "all", synthesis = FALSE)
addMonotones(mydata1)

}

}
