% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRamps.R
\name{addRamps}
\alias{addRamps}
\title{Ramp of an area}
\usage{
addRamps(x, ignoreMustRun = FALSE)
}
\arguments{
\item{x}{Object of class \code{antaresData} containing data for areas. it must
contain the column \code{BALANCE} and either the column "netLoad" or
the columns needed to compute the net load.}

\item{ignoreMustRun}{If \code{TRUE}, the production in must run mode is not substracted to the
net load.}
}
\value{
\code{addRamps} modify its input by adding to it three new columns:
\item{netLoadRamp}{Ramp of the net load of an area.}
\item{balanceRamp}{Ramp of the balance of an area.}
\item{areaRamp}{Sum of the two previous columns.}

For convenience the function invisibly returns the modified input.
}
\description{
This function computes the ramp of the consumption and the balance of areas
and/or districts and add them to the data.
}
\examples{
\dontrun{

  mydata <- readAntares(areas = "all", mustRun = TRUE, timeStep = "monthly")

  addRamps(mydata)
}

}

