% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExternalDependancies.R
\name{externalDependancies}
\alias{externalDependancies}
\title{External Dependancies in imports and exports}
\usage{
externalDependancies(x, timeStep = "annual", synthesis = FALSE)
}
\arguments{
\item{x}{An object created with function \code{\link[antaresRead]{readAntares}}. It
  must contain data for areas and/or distritcs . More specifically this
  function requires the columns \code{hstorPMaxAvg},
  and \code{netLoad}. To get these columns,
  one has to invoke \code{\link[antaresRead]{readAntares}} with the parameter \code{hydroStorageMaxPower = TRUE}
  and \code{\link[antaresProcessing]{addNetLoad}}
  (see examples).

  Moreover it needs to have a hourly time step.

  This object must also contain linkCapacity if there was virtual areas remove by \code{\link[antaresRead]{removeVirtualAreas}}
  to be able to calculate pumping and storage capacities.}

\item{timeStep}{Desired time step for the result.}

\item{synthesis}{If TRUE, average external dependncies are returned. Else the function returns external dependncies
per Monte-Carlo scenario.}
}
\value{
A data.table of class \code{antaresDataTable} with the following columns:
\item{area}{Area name.}
\item{timeId}{Time id and other time columns.}
\item{pumping}{capacity of pumping}
\item{storage}{capacity of storage}
\item{exportsLevel}{netLoad + pumping}
\item{importsLevel}{netLoad - `AVL DTG` - hydroStorageMaxPower - storage > 0}
\item{exportsFrequency}{number of time step where this criteria is satisfied\cr
  criteria : netLoad + pumping < 0
}
\item{importsFrequency}{number of time step where this criteria is satisfied\cr
  criteria : netLoad - `AVL DTG` - hydroStorageMaxPower - storage > 0
}
}
\description{
This function computes the depandance in imports and export for each area or districts at a
given time step. Depandance in imports represents moments where imports are required
to have no loss of load. Depandance in exports represents moments where exports are required to
have no spilled energy.
}
\examples{
\dontrun{
mydata <- readAntares(areas = "all",
                      hydroStorageMaxPower = TRUE)
externalDependancies(mydata)

mydata <- readAntares(districts = "all",
                      hydroStorageMaxPower = TRUE)
externalDependancies(mydata)

# if there are some virtual areas, include link and linkCapacity
mydata <- readAntares(areas = "all", link = "all", linkCapacity=TRUE,
                      hydroStorageMaxPower = TRUE)
removeVirtalAreas(mydata)
externalDependancies(mydata, ignoreMustRun = TRUE)

# Example that minimises the data imported
mydata <- readAntares(areas = "all", TODO = ...,
                      select = c("H. ROR", "WIND", "SOLAR", "MISC. NDG",
                                 "LOAD", "BALANCE"))

externalDependancies(mydata)
}

}

