% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{margins}
\alias{margins}
\title{Upward and downward margins for an area}
\usage{
margins(x, type = "upward", ignoreMustRun = FALSE, clusterDesc = NULL)
}
\arguments{
\item{x}{An object created with function \code{\link[antaresRead]{readAntares}}. It
must contain data for areas and for clusters. More specifically this
function requires the columns \code{hstorPMaxAvg},
and \code{mustRunPartial}. To get these columns,
one has to invoke \code{\link[antaresRead]{readAntares}} with parameters:
\code{mustRun = TRUE, hydroStorageMaxPower = TRUE}.
(see examples).}

\item{type}{Type of margins to compute. Possibilities : \cr
both = upward and downward margins are computed\cr
upward = only upward margins\cr
downward = only downward margins\cr}

\item{ignoreMustRun}{Should must run productions be ignored in the computation? It should be
\code{TRUE} only if the studied areas have no clusters in must run.}

\item{clusterDesc}{A table created with the function \code{\link[antaresRead]{readClusterDesc}}.
If is this parameter is set to \code{NULL} (the default), then the function
attemps to read the needed data in the same study as \code{x}.}
}
\value{
A data.table of class \code{antaresDataTable} with the following columns:
\item{area}{Area name.}
\item{timeId}{Time id and other time columns.}
\item{thermalAvailability}{Sum of thermal availabilities of all cluster of an area.\cr
                     formula = x$areas[, .(`AVL DTG`)]
}
\item{thermalPmin}{
  Sum of thermal minimum power of all clusters of an area. The minimum power
  of a cluster is defined as the maximum of the partial must run and the
  minimum stable power of the cluster. If \code{ignoreMustRun = TRUE}, it
  is simply equal to the minimum stable power of the cluster.\cr
                     formula = pmax(min.stable.power*NODU, mustRunTotal)
}
\item{pumping}{
}
\item{storage}{
}
\item{isolatedUpwardMargin}{\cr
  formula = `AVL DTG` + hstorPMaxAvg + storage + `H. ROR` + WIND + SOLAR + `MISC. NDG` - LOAD
}
\item{isolatedDownwardMargin}{\cr
  formula = thermalPmin - pumping + `H. ROR` + WIND + SOLAR + `MISC. NDG` - LOAD
}
\item{interconnectedUpwardMargin}{\cr
  formula = isolatedUpwardMargin - BALANCE
}
\item{interconnectedDownwardMargin}{\cr
  formula = isolatedDownwardMargin + BALANCE
}
}
\description{
This function computes the upward and downward margins for each area at a
given time step. Upward margin represents ... Downward margin represents ...
There are two versions of both indicators : isolated and interconnected.
The only difference is that the interconnected version takes into account
the balance of the area.
}
\examples{
\dontrun{
mydata <- readAntares(areas = "all", clusters = "all", mustRun = TRUE,
                      hydroStorageMaxPower = TRUE)
margins(mydata)

# If there is no must run in the study, use ignoreMustRun to avoid useless
# data importation and computations

mydata <- readAntares(areas = "all", clusters = "all",
                      hydroStorageMaxPower = TRUE)
margins(mydata, ignoreMustRun = TRUE)

# Example that minimises the data imported
mydata <- readAntares(areas = "all", clusters = "all",
                      hydroStorageMaxPower = TRUE,
                      select = c("H. ROR", "WIND", "SOLAR", "MISC. NDG",
                                 "LOAD", "BALANCE"))

margins(mydata, ignoreMustRun = TRUE)
}

}
