% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netLoad.R
\name{netLoad}
\alias{netLoad}
\title{Net load of areas}
\usage{
netLoad(x, timeStep = "hourly", ignoreMustRun = FALSE)
}
\arguments{
\item{x}{An \code{antaresData} object created with readAntares. unless
\code{ignoreMustRun} is true, it must
have a column \code{mustRunTotal}.}

\item{timeStep}{Desired time step for the result.}

\item{ignoreMustRun}{If \code{TRUE}, the production in must run mode is not substracted to the
net load.}
}
\value{
A data.table of class \code{antaresDataTable}. It contains one line per
area, timeId and scenario and has the following columns:
\item{area}{Name of the area}
\item{timeId}{time id and other time columns}
\item{mcYear}{id of the Monte-Carlo scenario}
\item{netLoad}{Net Load}
}
\description{
This function computes the net load of areas at the desired time step. Net
load is the load of an area minus productions that are not controlled:
wind, solar, hydraulic run of river, etc. the production of clusters in
must run mode is also subtracted by default.
}
\examples{
\dontrun{
mydata <- readAntares(areas = "all", synthesis = FALSE, mustRun= TRUE)

netLoad(mydata)
netLoad(mydata, "monthly")

}

}

