% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramps.R
\name{ramps}
\alias{ramps}
\title{Ramp of an area}
\usage{
ramps(x, ignoreMustRun = FALSE)
}
\arguments{
\item{x}{Object of class \code{antaresData} containing data for areas. It must
contain the columns needed to compute the net load and the column
\code{BALANCE}.}

\item{ignoreMustRun}{If \code{TRUE}, the production in must run mode is not substracted to the
net load.}
}
\value{
A data.table of class \code{antaresDataTable} containing the foloowing columns:
\item{area}{Area name.}
\item{timeId}{time id and other time columns depending on the time step of the data.}
\item{mcYear}{Scenario id. Visible only if it is present in the source data.}
\item{netLoadRamp}{Ramp of the net load of an area.}
\item{balanceRamp}{Ramp of the balance of an area.}
\item{areaRamp}{Sum of the two previous columns.}
}
\description{
This function computes the ramp of the consumption and the balance of an area
at a given time step.
}
\examples{
\dontrun{
  mydata <- readAntares(areas = "all", mustRun = TRUE)
  tamps(mydata)
}

}

