% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surplus.R
\name{surplus}
\alias{surplus}
\title{Compute economic surplus}
\usage{
surplus(x, timeStep = "annual", groupByDistrict = FALSE)
}
\arguments{
\item{x}{an object of class "antaresDataList" created with the function
\code{readAntares}. It has to contain some areas and all the links that are
connected to these areas. Moreover it needs to have a hourly time step.}

\item{timeStep}{Desired time step for the result.}
}
\value{
A data.table with the wollowing columns:
\item{area}{Name of the area.}
\item{timeId}{timeId and other time columns.}
\item{consumerSurplus}{The surplus of the consumers of some area.}
\item{producerSurplus}{The surplus of the producers of some area.}
\item{storageSurplus}{Surplus created by storage/flexibility areas.}
\item{congestionFees}{The congestion fees of a given area. It equals to half
  the congestion fees of the links connected to that area.}
\item{globalSurplus}{Sum of the consumer surplus, the producer surplus and
  the congestion fees.}
}
\description{
This function computes the economic surplus for the consumers, the producers
and the global surplus of an area.
}
\examples{
\dontrun{
mydata <- readAntares(areas = "all", links = "all")

mySurplus <- surplus(mydata)
}

}

