% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surplusClusters.R
\name{surplusClusters}
\alias{surplusClusters}
\title{Compute the surplus of clusters}
\usage{
surplusClusters(x, timeStep = "annual")
}
\arguments{
\item{x}{An \code{antaresData} object created with \code{readAntares}. It must
contain an element \code{clusters} and an element \code{areas} with at
least the column \code{MRG. PRICE}.}

\item{timeStep}{Desired time step for the result}
}
\value{
A data.table of class \code{antaresDataTable} with the following columns:
\item{area}{Area name.}
\item{cluster}{Cluster name.}
\item{timeId}{Time id and other time columns.}
\item{surplus}{Surplus of a given cluster at a given time step.}
}
\description{
This function computes the surplus of clusters of interest. The surplus of a
cluster is equal to its production times the marginal cost of the area it
belongs to minus variable, fixed and startup costs.
}
\examples{
\dontrun{
mydata <- readAntares(areas = "all", clusters = "all", select = "MRG. PRICE")

surplusClusters(mydata)

}

}

