% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermalGeneration.R
\name{termalGeneration}
\alias{termalGeneration}
\title{compute termal generation from study}
\usage{
termalGeneration(opts = simOptions())
}
\arguments{
\item{opts}{\code{simOptions} obtain wich \link[antaresRead]{setSimulationPath}}
}
\description{
compute termal generation from study
}
