% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/understandPrice.R
\name{understandThePrice}
\alias{understandThePrice}
\title{understandThePrice}
\usage{
understandThePrice(antaresData = NULL, opts = simOptions())
}
\arguments{
\item{antaresData}{Object of class \code{antaresData} created with function
\code{\link[antaresRead]{readAntares}}. antaresData must contains hourly detail data
for areas, clusters and links.
Areas data must contains "H. STOR".
Links data must contains transCapacityDirect and transCapacityIndirect.
Clusters data must contains minGenModulation.}
}
\value{
\code{understandPrice} modifies its input by adding to clusters
data columns "market.bid.cost" and "must.run".
}
\description{
This function gives some explanations about the price.
}
\examples{
\dontrun{
# First simulation
studyPath <- "path/to/study/"

setSimulationPath(studyPath, 1)
myData <- readAntares(showProgress = FALSE,
                      mcYears = "all",
                      areas = "all",
                      clusters = "all")
understandThePrice(antaresData = myData)

}

}
