% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCongestionLink.R
\name{addCongestionLink}
\alias{addCongestionLink}
\title{Load factors of link}
\usage{
addCongestionLink(x)
}
\arguments{
\item{x}{Object of class \code{antaresData} created with function
\code{\link[antaresRead]{readAntares}}. It must contain the columns
\code{CONG. PROB +} and \code{CONG. PROB -}.}
}
\value{
\code{addLoadFactorLink}
}
\description{
This function computes the load factor of link and add it to an
\code{antaresData} object.
}
\examples{
\dontrun{
# Data required by the function
showAliases("loadFactorLink")

mydata <- readAntares(select = "loadFactorLink")
addCongestionLink(mydata)
names(mydata)
}

}
