% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addExportAndImport.R
\name{addExportAndImport}
\alias{addExportAndImport}
\title{Export and import of areas or districts}
\usage{
addExportAndImport(x, addCapacities = FALSE)
}
\arguments{
\item{x}{an object of class "antaresDataList" created with the function
\code{readAntares}. It has to contain some areas and all the links that are
connected to these areas. Moreover the function \code{\link{removeVirtualAreas}} must be call before.}

\item{addCapacities}{If \code{TRUE}, export and import capacities are added.}
}
\description{
This function computes the export and import of areas or districts and add it to an
\code{antaresData} object.
}
\examples{
\dontrun{
mydata <- readAntares(areas = "all", links="all", districts ="all" , synthesis = FALSE)

mydata<-removeVirtualAreas(
  mydata,
  storageFlexibility = c(getAreas("psp"),getAreas("hub")),
  production = getAreas("offshore")
)

mydata<-addExportAndImport(mydata)

names(mydata$areas)
names(mydata$districts)

}
}

