% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addloadFactorLink.R
\name{addLoadFactorLink}
\alias{addLoadFactorLink}
\title{Load factors of link}
\usage{
addLoadFactorLink(x)
}
\arguments{
\item{x}{Object of class \code{antaresData} created with function
\code{\link[antaresRead]{readAntares}}. It must contain the columns
\code{transCapacityDirect} and \code{transCapacityIndirect}.}
}
\value{
\code{addLoadFactorLink} modifies its input by adding to it a column "loadFactor". For
convenience, it invisibly returns the modified input.
  loadFactor represent the proportion of
  the installed capacity of a link that is effectively used\cr
  Formula: `FLOW LIN` / transCapacity
}
\description{
This function computes the load factor of link and add it to an
\code{antaresData} object.
}
\examples{
\dontrun{
mydata <- readAntares(link = "all", linkCapacity=TRUE)

addLoadFactorLink(mydata)
names(mydata)

# Example that minimizes the quantity of data read
mydata <- readAntares(links = "all", timeStep = "annual",select = c("FLOW LIN."), linkCapacity=TRUE)
addLoadFactorLink(mydata)
}

}

