% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMonotone.R
\name{addMonotones}
\alias{addMonotones}
\title{addMonotones}
\usage{
addMonotones(antaresData = NULL, variable = NULL)
}
\arguments{
\item{antaresData}{Object of class \code{antaresData} created with function
\code{\link[antaresRead]{readAntares}}.}

\item{variable}{An ANTARES variable.}
}
\value{
\code{addMonotones} modifies its input by adding monotones.
}
\description{
This function compute monotone for some variables.
}
\examples{
\dontrun{
# First simulation
studyPath <- "path/to/study/"

setSimulationPath(studyPath, 1)
myData1 <- readAntares(areas = "all",
districts = "all", synthesis = FALSE)
addMonotones(antaresData = myData1,
variable = "LOAD")

}

}
