% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariableAntares.R
\name{getValues}
\alias{getValues}
\title{Get values of a variable}
\usage{
getValues(data = NULL, variable = NULL, mcYear = "all")
}
\arguments{
\item{data}{an object of class "antaresData" created with the function
\code{readAntares}.}

\item{variable}{a variable of data}

\item{mcYear}{set of mcYear}
}
\description{
Get all the values of a variable for some years Monte Carlo
}
\examples{
\dontrun{


mydata <- readAntares(areas="all", select="LOAD")
getVariableAntares(mydata, variable="LOAD")
names(mydata$areas)

}
}
