% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValues.R
\name{getValues}
\alias{getValues}
\title{Get values of a variable}
\usage{
getValues(data = NULL, variable = NULL, mcyear = "all")
}
\arguments{
\item{data}{an object of class "antaresData" created with the function
\code{readAntares}.}

\item{variable}{a variable of data}

\item{mcyear}{set of mcYear}
}
\description{
Get all the values of a variable for some years Monte Carlo
}
\examples{
\dontrun{

mydata <- readAntares(areas="all",clusters="all", select="LOAD")
getValues(mydata$areas, variable="LOAD")
getValues(myData$clusters, variable = "production")

}
}
