% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeAllTableByTime.R
\name{mergeAllAntaresData}
\alias{mergeAllAntaresData}
\title{Merge all antaresDataSets}
\usage{
mergeAllAntaresData(X)
}
\arguments{
\item{X}{antaresData}
}
\description{
Merge all antaresDataSets
}
\examples{
\dontrun{
setSimulationPath("Mystud", 1)
dta <- readAntares(areas = "all", links = "all", clusters = "all", districts = "all")
dta <- mergeAllAntaresData(dta)
}

}
