% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulation.R
\name{modulation}
\alias{modulation}
\title{Compute the modulation of cluster units}
\usage{
modulation(x, timeStep = "annual")
}
\arguments{
\item{x}{An \code{antaresData} object created with \code{readAntares}. It must
contain the hourly detailed results for clusters}

\item{timeStep}{Desired time step for the result}
}
\value{
A data.table of class \code{antaresDataTable} with the following columns:
\item{area}{Area name.}
\item{cluster}{Cluster name.}
\item{timeId}{Time id and other time columns.}
\item{meanUpWardModulation}{...}
\item{meanDownWardModulation}{...}
\item{meanAbsoluteModulation}{...}
\item{maxUpWardModulation}{...}
\item{maxDownWardModulation}{...}
\item{maxAbsoluteModulation}{...}
}
\description{
This function computes the modulation of cluster units.
}
\examples{
\dontrun{
mydata <- readAntares(clusters = "all")

modulation(mydata)

}

}

