% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surplusClusters.R
\name{surplusClusters}
\alias{surplusClusters}
\title{Compute the surplus of clusters}
\usage{
surplusClusters(x, timeStep = "annual", synthesis = FALSE,
  surplusLastUnit = FALSE, clusterDesc = NULL)
}
\arguments{
\item{x}{An \code{antaresData} object created with \code{readAntares}. It must
contain an element \code{clusters} and an element \code{areas} with at
least the column \code{MRG. PRICE}.}

\item{timeStep}{Desired time step for the result.}

\item{synthesis}{If TRUE, average surpluses are returned. Else the function returns surpluses
per Monte-Carlo scenario.}

\item{surplusLastUnit}{Should the surplus of the last unit of a cluster be computed ? If
\code{TRUE}, then \code{x} must have been created with the option
\code{thermalAvailabilities=TRUE} in order to contain the required column
"available units"}

\item{clusterDesc}{A table created with the function \code{\link[antaresRead]{readClusterDesc}}.
If is this parameter is set to \code{NULL} (the default), then the function
attemps to read the needed data in the same study as \code{x}.}
}
\value{
A data.table of class \code{antaresDataTable} with the following columns:
\item{area}{Area name.}
\item{cluster}{Cluster name.}
\item{timeId}{Time id and other time columns.}
\item{surplusPerUnit}{Average surplus per unit of the cluster.\cr
 formula = (`MRG. PRICE` * production - prodCost - startupCost) / unitcount }
\item{surplusLastUnit}{Surplus of the last unit of the cluster.\cr
formula = (`MRG. PRICE` * prodLastUnit - prodCost / pmax(1, NODU) - startup.cost)}
\item{totalSurplus}{Surplus of all units of the cluster.\cr
formula = `MRG. PRICE` * production - prodCost - startupCost}
\item{nbHoursGeneration}{It represents the production of a cluster expressed
in number of hours of production at the total capacity of the cluster. It is
equal to the production divided by the capacity of the cluster.\cr
formula = production / (unitcount * nominalcapacity)}
\item{economicGradient}{Economic gradient of a cluster. It is equal to
the surplus per unit divided by the capacity of a unit.\cr
formula = surplusPerUnit / nominalcapacity }
}
\description{
This function computes the surplus of clusters of interest. The surplus of a
cluster is equal to its production times the marginal cost of the area it
belongs to minus variable, fixed and startup costs.
}
\examples{
\dontrun{

mydata <- readAntares(areas = "all", clusters = "all", select = "MRG. PRICE",
                      synthesis = FALSE)
surplusClusters(mydata)

# Computing the surplus of the last unit of a cluster requires the additional
# column "availableUnits". To add this column, one has to use parameter
# "thermalAvailabilities = TRUE" in readAntares.

mydata <- readAntares(areas = "all", clusters = "all", select = "MRG. PRICE",
                      thermalAvailabilities = TRUE, synthesis = FALSE)
surplusClusters(mydata, surplusLastUnit = TRUE)

}

}

