% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surplusSectors.R
\name{surplusSectors}
\alias{surplusSectors}
\title{Compute the surplus of sectors}
\usage{
surplusSectors(x, sectors = c("thermal", "WIND", "SOLAR", "H. ROR",
  "H. STOR"), timeStep = "annual")
}
\arguments{
\item{x}{Object of class \code{antaresData} created with \code{readAntares}. It needs
to contain hourly detailed results of a simulation. Moreover, it must contain
area data and if thermal sectors are required, cluster data.}

\item{sectors}{vector containing the name of the sector wor which surplus needs to be
computed. Possible values are "thermal", "WIND", "SOLAR", "H. ROR", "H. STOR".
If it contains the value "thermal", then the parameter \code{x} has to contain
cluster data.}

\item{timeStep}{Desired time step for the result.}
}
\value{
A data.table of class "antaresData". Ir contains one column per sector
containing the surplus of that sector for a given area and timeId.
}
\description{
This function computes the surplus of sectors for each area and time step.
For sectors wind, solar, hydraulic storage and run of river, production costs
are assumed to be equal to 0.
}
\examples{
\dontrun{
mydata <- readAntares(areas = "all", clusters = "all")
surplusSectors(mydata)
}

}

