% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermalCapacities.R
\name{thermalGeneration}
\alias{thermalGeneration}
\title{compute thermal capacities from study}
\usage{
thermalGeneration(opts = simOptions())
}
\arguments{
\item{opts}{\code{simOptions} obtain wich \link[antaresRead]{setSimulationPath}}
}
\description{
compute thermal capacities from study
}
