% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermalGeneration.R
\name{thermalGroupCapacities}
\alias{thermalGroupCapacities}
\title{compute thermal capacities from study}
\usage{
thermalGroupCapacities(opts = simOptions())
}
\arguments{
\item{opts}{\code{simOptions} obtain which \link[antaresRead]{setSimulationPath}}
}
\description{
compute thermal capacities from study
}
