% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermaGrouplCapacities.R
\name{thermalGroupCapacities}
\alias{thermalGroupCapacities}
\title{compute thermal group capacities from study}
\usage{
thermalGroupCapacities(opts = simOptions())
}
\arguments{
\item{opts}{\code{simOptions} obtain wich \link[antaresRead]{setSimulationPath}}
}
\description{
compute thermal group capacities from study
}
